@@ -1,53 +1,58 @@
-source 'https://ruby.taobao.org'
+source 'https://gems.ruby-china.com'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.0'
+gem 'rails', '~> 5.0.1'
 # Use sqlite3 as the database for Active Record
 gem 'sqlite3'
+gem 'pg'
+# Use Puma as the app server
+gem 'puma', '~> 3.0'
 # Use SCSS for stylesheets
 gem 'sass-rails', '~> 5.0'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
-# See https://github.com/sstephenson/execjs#readme for more supported runtimes
+gem 'coffee-rails', '~> 4.2'
+# See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
+# Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
+gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
-
+gem 'jbuilder', '~> 2.5'
+# Use Redis adapter to run Action Cable in production
+# gem 'redis', '~> 3.0'
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
 
-# Use Unicorn as the app server
-# gem 'unicorn'
-
-# Use Capistrano for deploymentge
+# Use Capistrano for deployment
 # gem 'capistrano-rails', group: :development
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem 'byebug'
-
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
-
-  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-  gem 'spring'
+  gem 'byebug', platform: :mri
 end
 
-  gem "twitter-bootstrap-rails", '3.2.0'
-  gem "devise", '~> 3.4.1'
+group :development do
+  # Access an IRB console on exception pages or by using <%= console %> anywhere in the code.
+  gem 'web-console'
+  gem 'listen', '~> 3.0.5'
+  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'spring'
+  gem 'spring-watcher-listen', '~> 2.0.0'
+end
+
+gem "twitter-bootstrap-rails", '4.0.0'
+  # gem "twitter-bootstrap-rails", '~> 3.2.0'
+  gem "devise"
+  # gem "devise", '~> 3.4.1'
   gem "cancancan"
   gem "rolify"
-  gem 'rails-i18n', '~> 4.0.0'
+  gem 'rails-i18n'
+  # gem 'rails-i18n', '~> 4.0.0'
   gem 'devise-i18n'
 
   group :development do
@@ -57,17 +62,20 @@ end
   gem 'jquery-ui-rails'
   gem 'rails-jquery-autocomplete'
 
-  gem 'seed-fu', '~> 2.3'
+  gem 'seed-fu'
+  # gem 'seed-fu', '~> 2.3'
   gem 'kaminari'
-  gem 'paperclip', "~> 4.2"
-  gem 'paranoia', "~> 2.0"
+  gem 'paperclip'
+  # gem 'paperclip', "~> 4.2"
+  gem 'paranoia'
+  # gem 'paranoia', "~> 2.0"
   gem 'smarter_csv'
   gem 'csv_builder'
-  gem 'ckeditor', '4.1.5'
+  gem 'ckeditor'
+  # gem 'ckeditor', '4.1.5'
   gem 'tinymce-rails'
   gem 'tinymce-rails-langs'
   gem 'mathjax-rails'
-  gem 'pg'
   gem 'sidekiq'
 
 # Use Capistrano for deployment
