@@ -1,4 +1,4 @@
-class Agreement < ActiveRecord::Base
+class Agreement < ApplicationRecord
   belongs_to :user
   belongs_to :comment
   acts_as_paranoid
