@@ -1,4 +1,4 @@
-class Classgroupscore < ActiveRecord::Base
+class Classgroupscore < ApplicationRecord
   belongs_to :user
   belongs_to :team
   has_many :classpersonscores, dependent: :destroy
