@@ -1,4 +1,4 @@
-class Lesson < ActiveRecord::Base
+class Lesson < ApplicationRecord
   resourcify
   belongs_to :user
   has_many :catalogs
