@@ -1,4 +1,4 @@
-class PhoneticNotation < ActiveRecord::Base
+class PhoneticNotation < ApplicationRecord
   belongs_to :phonetic
   belongs_to :word
 end
