@@ -1,4 +1,4 @@
-class QuizItem < ActiveRecord::Base
+class QuizItem < ApplicationRecord
   belongs_to :user
   belongs_to :quiz
   belongs_to :practice
