@@ -1,4 +1,4 @@
-class WordParser < ActiveRecord::Base
+class WordParser < ApplicationRecord
   belongs_to :word
   belongs_to :lesson
   belongs_to :sentence
