@@ -1,4 +1,4 @@
-class WordsReport < ActiveRecord::Base
+class WordsReport < ApplicationRecord
   belongs_to :lesson
   acts_as_paranoid
   validates :lesson_id, :md,  presence: true
