@@ -60,6 +60,7 @@ class Users::SessionsController < Devise::SessionsController
 
   # You can put the params you want to permit in the empty array.
   def configure_sign_in_params
-    devise_parameter_sanitizer.for(:sign_in) << :attribute
+    # devise_parameter_sanitizer.for(:sign_in) << :attribute
+    devise_parameter_sanitizer.permit(:sign_in, keys: [:user, :password])
   end
 end
